+++ packages/nuxt-module/playground/nuxt.config.ts
@@ -1,5 +1,0 @@
-// https://nuxt.com/docs/api/configuration/nuxt-config
-export default defineNuxtConfig({
-  modules: ['nuxt-simple-ui-vue'],
-  devtools: { enabled: true }
-})