+++ packages/docs/CHANGELOG.md
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [350f81f]
-  - simple-vue@1.1.0
+  - simple-daisy-vue@1.1.0